Command options
lil bit refactoring
Added options to the command for greater flexibility

+++ converter.py
@@ -3,15 +3,22 @@ import re
 from os import path
 from pathlib import Path
 from PIL import Image
+import sys
 
-DEFAULT_SRC_FOLDER = Path("HTMLPages")
-PATH_TO_IMAGES = Path("HTMLPages/images/")
+DEFAULT_DST_FOLDER = Path("HTMLPages")
+PATH_TO_IMAGES = Path("images/")
 ROOT_HTML = "index.html"
 ROOT_CSS = "style.css"
-ROOT_CSS_PATH = DEFAULT_SRC_FOLDER / ROOT_CSS
-ROOT_HTML_PATH = DEFAULT_SRC_FOLDER / ROOT_HTML
 MENU_PLACEHOLDER_NAME = "<div class='PLACEHODER'></div>"
 
+ARGS = {
+    "help": "--help",
+    "src": "--src",
+    "dst": "--dst",
+    "overwrite": "--overwrite",
+    "root-index": "--ri",
+    "root-css": "--rc"
+}
 
 def set_menu(base_url: Path, root: Path) -> str:
     html = "<ul>\n"
@@ -43,7 +50,7 @@ def create_menu(root_path: Path):
         if any(part.startswith('.') for part in path.parts):
             continue
         if not path.is_dir():
-            print(path.name)
+            # print(path.name)
             if path.suffix == ".html":
                 with open(path, "r", encoding="utf-8") as html:
                     data = html.read()
@@ -55,19 +62,19 @@ def create_menu(root_path: Path):
 def get_relative_path(src: Path, dst: Path):
     return path.relpath(dst, src)
 
-def create_css():
+def create_css(dst: Path):
     with open(ROOT_CSS, "r") as css:
         css_text = css.read()
-    with open(ROOT_CSS_PATH, "w") as f:
+    with open(dst / ROOT_CSS, "w") as f:
         f.write(css_text)
 
-def create_root():
+def create_root(dst: Path):
     with open(ROOT_HTML, "r") as index:
         text = index.read()
-    with open(ROOT_HTML_PATH, "w") as index:
+    with open(dst / ROOT_HTML, "w") as index:
         index.write(text)
 
-def copy_directory(src_path: Path, dst_path: Path):
+def copy_directory(src_path: Path, dst_path: Path, OVERWRITE: bool):
 
     image_extensions = [
         ".jpg",
@@ -100,7 +107,7 @@ def copy_directory(src_path: Path, dst_path: Path):
                 text = f.read()
 
             blockquote = preprocess_callouts(text)
-            photos = set_photos(blockquote, target)
+            photos = set_photos(blockquote, target, dst_path)
 
             md_to_html = markdown.markdown(photos,extensions=["fenced_code"])
 
@@ -110,8 +117,8 @@ def copy_directory(src_path: Path, dst_path: Path):
                 <html lang="uk">
                     <head>
                         <meta charset="UTF-8">
-                        <title>Document</title>
-                        <link rel="stylesheet" href={get_relative_path(target, ROOT_CSS_PATH)}>
+                        <title>{target.stem}</title>
+                        <link rel="stylesheet" href={get_relative_path(target.parent, dst_path / ROOT_CSS)}>
                     </head>
                     <body>
                         <h1 class="topic-name">{target.stem}</h2>
@@ -124,19 +131,21 @@ def copy_directory(src_path: Path, dst_path: Path):
                     </body>
                 </html>
                 """
-
-            with open(target, "w", encoding="utf-8") as f:
-                f.write(html_content)
+            if not target.exists() or (target.exists() and OVERWRITE):
+                print(target.name)
+                with open(target, "w", encoding="utf-8") as f:
+                    f.write(html_content)
         elif path.suffix in image_extensions:
             img = Image.open(path)
             img.save(target)
 
-def set_photos(html: str, current_path: Path) -> str:
+def set_photos(html: str, current_path: Path, dst_path: Path) -> str:
     pattern = re.compile(r'!\[\[(.+?)\]\]')
 
     def replacer(match):
         filename = match.group(1).strip()
-        image_path = path.join(get_relative_path(path.dirname(current_path), PATH_TO_IMAGES), filename)
+        image_path = path.join(get_relative_path(current_path.parent, dst_path / PATH_TO_IMAGES), filename)
+        
         return f'<div class="image-div"><img src="{image_path}" alt="{filename}" /></div>'
 
     return pattern.sub(replacer, html)
@@ -164,10 +173,60 @@ def preprocess_callouts(md_text: str) -> str:
 
     return "\n".join(processed_lines)
 
-if __name__ == "__main__":
+#-------------------HELP---------------
 
-    copy_directory(Path("D:\Web\Web"), Path(DEFAULT_SRC_FOLDER))
-    create_css()
-    create_root()
-    create_menu(Path(DEFAULT_SRC_FOLDER))
+def print_help():
+    print(ARGS)
+
+#------------------MAIN----------------
+
+if __name__ == "__main__":
+    CREATE_CSS = False
+    CREATE_HTML = False
+    OVERWRITE = False
+    arguments = []
+    for arg in sys.argv[1:]:
+        arguments.append(arg)
+    if len(arguments) == 0 or (len(arguments) == 1 and arguments[0] != "--help"):
+        print_help()
+    elif not ARGS["src"] in arguments: 
+        print("Please, provide the obssdian vault with '--src VAULT'")
+    else:
+        i = 0
+        for arg in arguments:
+            if arg == ARGS["help"]: 
+                print_help()
+                sys.exit()
+            elif arg == ARGS["src"]:
+                try:
+                    src_vault = Path(arguments[i+1])
+                    if not src_vault.exists():
+                        print(f"Provided source path does not exist: {src_vault}")
+                        sys.exit()
+                except:
+                    print_help()
+                    sys.exit()
+            elif arg == ARGS["dst"]:
+                try:
+                    DEFAULT_DST_FOLDER = Path(arguments[i+1])
+                except:
+                    print(f"Path is invalid: {src_vault}")
+                    print_help()
+                    sys.exit()
+            elif arg == ARGS["root-index"]:
+                CREATE_HTML=True
+            elif arg == ARGS["root-css"]:
+                CREATE_CSS=True
+            elif arg == ARGS["overwrite"]:
+                OVERWRITE=True
+            
+            i+=1
+        
+        copy_directory(Path(src_vault), Path(DEFAULT_DST_FOLDER), OVERWRITE)
+        if CREATE_CSS:
+            create_css(DEFAULT_DST_FOLDER)
+        if CREATE_HTML:
+            create_root(DEFAULT_DST_FOLDER)
+        create_menu(Path(DEFAULT_DST_FOLDER))
+        print(f"Vault was converted to {DEFAULT_DST_FOLDER}")
